Rename tool-call -> tool-calls in frontend
