+++ apps/frontend/src/components/chat-messages.tsx
@@ -2,8 +2,8 @@ import { Streamdown } from 'streamdown';
 import { useEffect, useMemo, useRef } from 'react';
 import { useParams, useRouterState } from '@tanstack/react-router';
 import { useStickToBottomContext } from 'use-stick-to-bottom';
-import { ToolCall } from './tool-call';
-import { ToolCallsGroup } from './tool-call/tool-calls-group';
+import { ToolCall } from './tool-calls';
+import { ToolCallsGroup } from './tool-calls/tool-calls-group';
 import { ReasoningAccordion } from './chat-message-reasoning-accordion';
 import { TextShimmer } from './ui/text-shimmer';
 import { MessageActions } from './chat-message-actions';
