+++ apps/frontend/src/components/tool-calls/display-chart-range-selector.tsx
