+++ apps/frontend/src/components/tool-calls/display-chart-series-toggle.tsx
