+++ apps/frontend/src/components/tool-calls/display-chart.tsx
