+++ apps/frontend/src/components/tool-calls/tool-calls-group.tsx
